chore: refactor editor states
