+++ editor/grida-react-canvas/index.ts
@@ -8,7 +8,5 @@ export {
   useTemplateDefinition,
 } from "./provider";
 export { StandaloneSceneContent } from "./renderer";
-export * from "./state";
 export * from "./viewport";
 export { default as standaloneDocumentReducer } from "./reducers";
-export type { Action as CanvasAction } from "./action";
