+++ editor/grida-react-canvas/reducers/methods/selection.ts
@@ -1,13 +1,12 @@
 import type { Draft } from "immer";
-import type { IDocumentEditorState } from "../../state";
-import { document } from "../../document-query";
+import { editor } from "@/grida-canvas";
 import assert from "assert";
 
 /**
  * TODO:
  * - validate the selection by config (which does not exists yet), to only select subset of children or a container, but not both. - when both container and children are selected, when transform, it will transform both, resulting in a weird behavior.
  */
-export function self_selectNode<S extends IDocumentEditorState>(
+export function self_selectNode<S extends editor.state.IEditorState>(
   draft: Draft<S>,
   mode: "reset" | "add" | "toggle",
   ...node_ids: string[]
@@ -15,7 +14,7 @@ export function self_selectNode<S extends IDocumentEditorState>(
   for (const node_id of node_ids) {
     assert(node_id, "Node ID must be provided");
     assert(
-      document.__getNodeById(draft, node_id),
+      editor.dq.__getNodeById(draft, node_id),
       `Node not found with id: "${node_id}"`
     );
   }
@@ -23,7 +22,7 @@ export function self_selectNode<S extends IDocumentEditorState>(
   switch (mode) {
     case "add": {
       const set = new Set([...draft.selection, ...node_ids]);
-      const pruned = document.pruneNestedNodes(
+      const pruned = editor.dq.pruneNestedNodes(
         draft.document_ctx,
         Array.from(set)
       );
@@ -39,7 +38,7 @@ export function self_selectNode<S extends IDocumentEditorState>(
           set.add(node_id);
         }
       }
-      const pruned = document.pruneNestedNodes(
+      const pruned = editor.dq.pruneNestedNodes(
         draft.document_ctx,
         Array.from(set)
       );
@@ -49,7 +48,7 @@ export function self_selectNode<S extends IDocumentEditorState>(
     case "reset": {
       // only apply if actually changed
       if (JSON.stringify(node_ids) !== JSON.stringify(draft.selection)) {
-        const pruned = document.pruneNestedNodes(draft.document_ctx, node_ids);
+        const pruned = editor.dq.pruneNestedNodes(draft.document_ctx, node_ids);
         draft.selection = pruned;
 
         // reset the active duplication as selection changed. see ActiveDuplication's note
@@ -61,7 +60,7 @@ export function self_selectNode<S extends IDocumentEditorState>(
   return draft;
 }
 
-export function self_clearSelection<S extends IDocumentEditorState>(
+export function self_clearSelection<S extends editor.state.IEditorState>(
   draft: Draft<S>
 ) {
   draft.selection = [];
