+++ editor/grida-react-canvas/reducers/methods/transform.ts
@@ -1,14 +1,10 @@
 import type { Draft } from "immer";
-import {
-  DEFAULT_SNAP_MOVEMNT_THRESHOLD_FACTOR,
-  type IDocumentEditorState,
-} from "../../state";
+import { editor } from "@/grida-canvas";
 import { self_insertSubDocument } from "./insert";
 import { self_try_remove_node } from "./delete";
-import { document } from "../../document-query";
 import { cmath } from "@grida/cmath";
 import { dnd } from "@grida/cmath/_dnd";
-import { domapi } from "../../domapi";
+import { domapi } from "../../../grida-canvas/backends/dom";
 import {
   getSnapTargets,
   snapObjectsTranslation,
@@ -41,9 +37,9 @@ const cardinal_direction_vector = {
   w: [-1, 0] as cmath.Vector2,
 } as const;
 
-export function self_update_gesture_transform<S extends IDocumentEditorState>(
-  draft: Draft<S>
-) {
+export function self_update_gesture_transform<
+  S extends editor.state.IEditorState,
+>(draft: Draft<S>) {
   if (draft.gesture.type === "idle") return;
   if (draft.gesture.type === "draw") return;
   if (draft.gesture.type === "corner-radius") return;
@@ -75,7 +71,7 @@ export function self_update_gesture_transform<S extends IDocumentEditorState>(
 }
 
 function __self_update_gesture_transform_translate(
-  draft: Draft<IDocumentEditorState>
+  draft: Draft<editor.state.IEditorState>
 ) {
   assert(draft.gesture.type === "translate", "Gesture type must be translate");
   assert(draft.scene_id, "scene_id is not set");
@@ -108,7 +104,7 @@ function __self_update_gesture_transform_translate(
       const to_be_cloned = initial_selection.slice();
 
       to_be_cloned.forEach((original_id, i) => {
-        const initial_parent_id = document.getParentId(
+        const initial_parent_id = editor.dq.getParentId(
           initial_snapshot.document_ctx,
           original_id
         );
@@ -194,7 +190,7 @@ function __self_update_gesture_transform_translate(
   switch (translate_with_hierarchy_change) {
     case "on": {
       // check if the cursor finds a new parent (if it escapes the current parent or enters a new parent)
-      const hits = draft.surface_raycast_detected_node_ids.slice();
+      const hits = draft.hits.slice();
 
       // filter out the...
       // 1. current selection (both original and cloned) and children of the current selection, recursive (both original and cloned)
@@ -206,13 +202,13 @@ function __self_update_gesture_transform_translate(
         ...initial_selection,
         ...initial_selection
           .map((node_id) =>
-            document.getChildren(draft.document_ctx, node_id, true)
+            editor.dq.getChildren(draft.document_ctx, node_id, true)
           )
           .flat(),
         ...initial_clone_ids,
         ...initial_clone_ids
           .map((node_id) =>
-            document.getChildren(draft.document_ctx, node_id, true)
+            editor.dq.getChildren(draft.document_ctx, node_id, true)
           )
           .flat(),
       ];
@@ -222,7 +218,7 @@ function __self_update_gesture_transform_translate(
         // [1]
         if (hierarchy_ids.includes(node_id)) return false;
 
-        const node = document.__getNodeById(draft, node_id);
+        const node = editor.dq.__getNodeById(draft, node_id);
         // [2]
         if (node.type !== "container") return false;
 
@@ -237,7 +233,7 @@ function __self_update_gesture_transform_translate(
       // update the parent of the current selection
       current_selection.forEach((node_id) => {
         //
-        const prev_parent_id = document.getParentId(
+        const prev_parent_id = editor.dq.getParentId(
           draft.document_ctx,
           node_id
         );
@@ -247,7 +243,7 @@ function __self_update_gesture_transform_translate(
 
         // unregister the node from the previous parent
         if (prev_parent_id) {
-          const parent = document.__getNodeById(
+          const parent = editor.dq.__getNodeById(
             draft,
             prev_parent_id
           ) as grida.program.nodes.i.IChildrenReference;
@@ -260,7 +256,7 @@ function __self_update_gesture_transform_translate(
 
         // register the node to the new parent
         if (new_parent_id) {
-          const new_parent = document.__getNodeById(
+          const new_parent = editor.dq.__getNodeById(
             draft,
             new_parent_id
           ) as grida.program.nodes.i.IChildrenReference;
@@ -272,7 +268,7 @@ function __self_update_gesture_transform_translate(
         }
 
         // update the context
-        draft.document_ctx = document.Context.from(draft.document).snapshot();
+        draft.document_ctx = editor.dq.Context.from(draft.document).snapshot();
       });
 
       if (is_parent_changed) {
@@ -305,7 +301,10 @@ function __self_update_gesture_transform_translate(
       guides: draft.ruler === "on" ? scene.guides : undefined,
     },
     adj_movement,
-    threshold(DEFAULT_SNAP_MOVEMNT_THRESHOLD_FACTOR, draft.transform)
+    threshold(
+      editor.config.DEFAULT_SNAP_MOVEMNT_THRESHOLD_FACTOR,
+      draft.transform
+    )
   );
 
   draft.surface_snapping = snapping;
@@ -314,10 +313,10 @@ function __self_update_gesture_transform_translate(
     let i = 0;
 
     for (const node_id of current_selection) {
-      const node = document.__getNodeById(draft, node_id);
+      const node = editor.dq.__getNodeById(draft, node_id);
       const r = translated[i++];
 
-      const parent_id = document.getParentId(draft.document_ctx, node_id);
+      const parent_id = editor.dq.getParentId(draft.document_ctx, node_id);
 
       let relative_position: cmath.Vector2;
       if (parent_id) {
@@ -365,7 +364,7 @@ function __self_update_gesture_transform_translate(
 }
 
 function __self_update_gesture_transform_translate_sort(
-  draft: Draft<IDocumentEditorState>
+  draft: Draft<editor.state.IEditorState>
 ) {
   assert(draft.gesture.type === "sort", "Gesture type must be translate-swap");
 
@@ -375,7 +374,7 @@ function __self_update_gesture_transform_translate_sort(
   // [moving node]
   // apply movement as-is to moving node
   const moving_rect = cmath.rect.translate(node_initial_rect, movement);
-  const moving_node = document.__getNodeById(
+  const moving_node = editor.dq.__getNodeById(
     draft,
     node_id
   ) as grida.program.nodes.i.IPositioning;
@@ -430,7 +429,7 @@ function __self_update_gesture_transform_translate_sort(
   // update the position of the real nodes (except the moving node)
   layout.objects.forEach((obj, i) => {
     if (obj.id === node_id) return;
-    const node = document.__getNodeById(
+    const node = editor.dq.__getNodeById(
       draft,
       obj.id
     ) as grida.program.nodes.i.IPositioning;
@@ -440,7 +439,7 @@ function __self_update_gesture_transform_translate_sort(
 }
 
 function __self_update_gesture_transform_scale(
-  draft: Draft<IDocumentEditorState>
+  draft: Draft<editor.state.IEditorState>
 ) {
   assert(draft.gesture.type === "scale", "Gesture type must be scale");
   assert(draft.scene_id, "scene_id is not set");
@@ -529,7 +528,7 @@ function __self_update_gesture_transform_scale(
       });
     } else {
       const cdom = new domapi.CanvasDOM(draft.transform);
-      const parent_id = document.getParentId(draft.document_ctx, node_id)!;
+      const parent_id = editor.dq.getParentId(draft.document_ctx, node_id)!;
       const parent_rect = cdom.getNodeBoundingRect(parent_id)!;
 
       assert(
@@ -582,7 +581,7 @@ function __self_update_gesture_transform_scale(
 }
 
 function __self_update_gesture_transform_rotate(
-  draft: Draft<IDocumentEditorState>
+  draft: Draft<editor.state.IEditorState>
 ) {
   assert(draft.gesture.type === "rotate", "Gesture type must be rotate");
   const { movement, selection } = draft.gesture;
@@ -601,7 +600,7 @@ function __self_update_gesture_transform_rotate(
   const q = Math.max(0.01, _user_q);
   const angle = cmath.quantize(_angle, q);
 
-  const node = document.__getNodeById(draft, selection);
+  const node = editor.dq.__getNodeById(draft, selection);
 
   draft.gesture.rotation = angle;
   draft.document.nodes[selection] = nodeReducer(node, {
