+++ editor/grida-react-canvas/reducers/node.reducer.ts
@@ -1,6 +1,6 @@
 import { produce, type Draft } from "immer";
 import grida from "@grida/schema";
-import type { NodeChangeAction } from "../action";
+import type { NodeChangeAction } from "../../grida-canvas/action";
 import type cg from "@grida/cg";
 import { v4 } from "uuid";
 import assert from "assert";
