+++ editor/grida-react-canvas/renderer.tsx
@@ -3,7 +3,7 @@
 import React, { useContext, useLayoutEffect, useMemo, useRef } from "react";
 import { useCurrentScene, useTransform } from "./provider";
 import { NodeElement } from "./nodes/node";
-import { domapi } from "./domapi";
+import { domapi } from "../grida-canvas/backends/dom";
 import { cmath } from "@grida/cmath";
 import { TransparencyGrid } from "@grida/transparency-grid/react";
 import { useMeasure } from "@uidotdev/usehooks";
