+++ editor/grida-react-canvas/viewport/surface-hooks.ts
@@ -1,15 +1,16 @@
 import React, { useLayoutEffect, useMemo, useRef, useState } from "react";
 import { useDocument, useNode } from "@/grida-react-canvas";
 import { analyzeDistribution } from "./ui/distribution";
-import { domapi } from "@/grida-react-canvas/domapi";
-import { document } from "@/grida-react-canvas/document-query";
+import { domapi } from "@/grida-canvas/backends/dom";
 import { rectToSurfaceSpace } from "@/grida-react-canvas/utils/transform";
 import { cmath } from "@grida/cmath";
 import type { ObjectsDistributionAnalysis } from "./ui/distribution";
 import grida from "@grida/schema";
 import { NodeWithMeta } from "../provider";
-import "core-js/features/object/group-by";
 import { is_direct_component_consumer } from "@/grida-canvas-utils/utils/supports";
+import { editor } from "@/grida-canvas";
+import "core-js/features/object/group-by";
+
 export interface SurfaceNodeObject {
   id: string;
   boundingRect: cmath.Rectangle;
@@ -255,7 +256,7 @@ export function useSelectionGroups(
       .filter((n) => n && n.active);
     return Object.groupBy(
       activenodes,
-      (it) => document.getParentId(state.document_ctx, it.id) ?? ""
+      (it) => editor.dq.getParentId(state.document_ctx, it.id) ?? ""
     );
   }, [state.document.nodes, state.document_ctx, __node_ids]);
 
@@ -366,7 +367,7 @@ export function useSingleSelection(
       const container = node as grida.program.nodes.ContainerNode;
       const { direction, mainAxisGap, crossAxisGap } = container;
       const axis = direction === "horizontal" ? "x" : "y";
-      const children = document.getChildren(state.document_ctx, node_id);
+      const children = editor.dq.getChildren(state.document_ctx, node_id);
       const children_rects = children
         .map((id) => cdom.getNodeBoundingRect(id))
         .filter((it): it is cmath.Rectangle => !!it);
