+++ editor/scaffolds/editor/action.ts
@@ -20,7 +20,7 @@ import type {
 } from "./state";
 import type { tokens } from "@grida/tokens";
 import type { DataQueryAction } from "../data-query";
-import type { CanvasAction } from "@/grida-react-canvas";
+import type { Action as CanvasAction } from "@/grida-canvas";
 import type { DataFormat } from "../data-format";
 import type { Platform } from "@/lib/platform";
 import grida from "@grida/schema";
