+++ editor/scaffolds/editor/reducer.ts
@@ -42,10 +42,7 @@ import datagridQueryReducer from "../data-query/data-query.reducer";
 import builderReducer from "@/grida-react-canvas/reducers";
 import assert from "assert";
 import grida from "@grida/schema";
-import {
-  IDocumentEditorState,
-  initDocumentEditorState,
-} from "@/grida-react-canvas/state";
+import { editor } from "@/grida-canvas";
 
 export function reducer(state: EditorState, action: EditorAction): EditorState {
   switch (action.type) {
@@ -119,7 +116,7 @@ export function reducer(state: EditorState, action: EditorAction): EditorState {
         document.state = builderReducer(
           document.state,
           _action
-        ) as IDocumentEditorState & { template_id: string };
+        ) as editor.state.IEditorState & { template_id: string };
       });
     }
 
@@ -379,7 +376,7 @@ export function reducer(state: EditorState, action: EditorAction): EditorState {
           __schema_valid: true,
           state: {
             template_id: template_id,
-            ...initDocumentEditorState({
+            ...editor.state.init({
               editable: true,
               debug: false,
               document: {
